temp

+++ Assets/ObjectLight/SpawnShadow/SpawnShadow.shader
@@ -92,7 +92,7 @@
                 return o;
             }
             
-            float4 frag(v2f i): SV_Target
+            half4 frag(v2f i): SV_Target
             {
                 SHADOW_CASTER_FRAGMENT(i)
             }
